+++ priv/templates/drab/drab.waiter.js
@@ -1,14 +1,19 @@
 Drab.on_connect(function(resp, drab) {
-  drab.channel.on("register_waiter", function(message) {
+  drab.channel.on("register_waiters", function(message) {
     // Drab.Query.update(attr: "drab_waiter_token", set: waiter_token, on: selector)
-    $(message.selector).attr("drab_waiter_token", message.drab_waiter_token)
-    $(message.selector).on(message.event_name, function(event) {
-      var t = $(this)
-      drab.channel.push("waiter", {
-        drab_waiter_token: message.drab_waiter_token, 
-        sender: payload(t, event)
+
+    message.waiters.forEach(function(waiter) {
+    $(waiter.selector).attr("drab_waiter_token", waiter.drab_waiter_token)
+      $(waiter.selector).off(waiter.event_name).on(waiter.event_name, function(event) {
+        var t = $(this)
+        drab.channel.push("waiter", {
+          drab_waiter_token: waiter.drab_waiter_token, 
+          sender: payload(t, event)
+        })
       })
     })
-
+  })
+  drab.channel.on("unregister_waiters", function(message) {
+    $(message.selector).off(message.event_name)
   })
 })